ramlserver

+++ JumpScale9/fs/SystemFS.py
@@ -45,25 +45,30 @@ def path_check(**arguments):
     """
     for argument, validators in arguments.items():
         if not isinstance(validators, set):
-            raise ValueError("Expected tuple of validators for argument %s" % argument)
+            raise ValueError(
+                "Expected tuple of validators for argument %s" % argument)
         for validator in validators:
             if validator not in {"required", "exists", "file", "dir", "pureFile", "pureDir"}:
-                raise ValueError("Unsupported validator '%s' for argument %s" % (validator, argument))
+                raise ValueError(
+                    "Unsupported validator '%s' for argument %s" % (validator, argument))
 
     def decorator(func):
         signature = inspect.signature(func)
         for argument in arguments:
             if signature.parameters.get(argument) is None:
-                raise ValueError("Argument %s not found in function declaration of %s" % (argument, func.__name__))
+                raise ValueError("Argument %s not found in function declaration of %s" % (
+                    argument, func.__name__))
 
         @wraps(func)
         def wrapper(*args, **kwargs):
-            jslocation = lambda: "%s." % args[0].__jslocation__ if hasattr(args[0], "__jslocation__") else ""
+            def jslocation(): return "%s." % args[0].__jslocation__ if hasattr(
+                args[0], "__jslocation__") else ""
             args = list(args)
             position = 0
             for parameter in signature.parameters.values():
                 if parameter.name in arguments:
-                    value = args[position] if position < len(args) else kwargs[parameter.name]
+                    value = args[position] if position < len(
+                        args) else kwargs[parameter.name]
                     if value and value.startswith("~%s" % os.sep):
                         if "HOME" in os.environ:
                             value = os.path.expanduser(value)
@@ -75,17 +80,23 @@ def path_check(**arguments):
                             kwargs[parameter.name] = value
                     validators = arguments[parameter.name]
                     if value and validators.intersection({"exists", "file", "dir", "pureFile", "pureDir"}) and not os.path.exists(value):
-                        raise ValueError("Argument %s in %s%s expects an existing path value! %s does not exist." % (parameter.name, jslocation(), func.__name__, value))
+                        raise ValueError("Argument %s in %s%s expects an existing path value! %s does not exist." % (
+                            parameter.name, jslocation(), func.__name__, value))
                     if "required" in validators and (value is None or value.strip() == ""):
-                        raise ValueError("Argument %s in %s%s should not be None or empty string!" % (parameter.name, jslocation(), func.__name__))
+                        raise ValueError("Argument %s in %s%s should not be None or empty string!" % (
+                            parameter.name, jslocation(), func.__name__))
                     if value and validators.intersection({"file", "pureFile"}) and not os.path.isfile(value):
-                        raise ValueError("Argument %s in %s%s expects a file path! %s is not a file." % (parameter.name, jslocation(), func.__name__, value))
-                    if value and "pureFile" in validators and os.path.islink(value) :
-                        raise ValueError("Argument %s in %s%s expects a file path! %s is not a file but a link." % (parameter.name, jslocation(), func.__name__, value))
+                        raise ValueError("Argument %s in %s%s expects a file path! %s is not a file." % (
+                            parameter.name, jslocation(), func.__name__, value))
+                    if value and "pureFile" in validators and os.path.islink(value):
+                        raise ValueError("Argument %s in %s%s expects a file path! %s is not a file but a link." % (
+                            parameter.name, jslocation(), func.__name__, value))
                     if value and validators.intersection({"dir", "pureDir"}) and not os.path.isdir(value):
-                        raise ValueError("Argument %s in %s%s expects a directory path! %s is not a directory." % (parameter.name, jslocation(), func.__name__, value))
+                        raise ValueError("Argument %s in %s%s expects a directory path! %s is not a directory." % (
+                            parameter.name, jslocation(), func.__name__, value))
                     if value and "pureDir" in validators and os.path.islink(value):
-                        raise ValueError("Argument %s in %s%s expects a directory path! %s is not a directory but a link." % (parameter.name, jslocation(), func.__name__, value))
+                        raise ValueError("Argument %s in %s%s expects a directory path! %s is not a directory but a link." % (
+                            parameter.name, jslocation(), func.__name__, value))
                 position += 1
             return func(*args, **kwargs)
         return wrapper
@@ -166,26 +177,32 @@ class SystemFS:
         """Remove a File
         @param path: string (File path required to be removed)
         """
-        self.logger.debug('Remove file with path: %s' % path)
-        if len(path) > 0 and path[-1] == os.sep:
-            path = path[:-1]
-        if os.path.islink(path):
-            os.unlink(path)
-        if self.exists(path):
-            os.remove(path)
-            self.logger.debug('Done removing file with path: %s' % path)
+        if not self.exists(path):
+            return
+        if self.isFile(path) or self.isLink(path):
+            self.logger.debug('Remove file with path: %s' % path)
+            if len(path) > 0 and path[-1] == os.sep:
+                path = path[:-1]
+            if os.path.islink(path):
+                os.unlink(path)
+            if self.exists(path):
+                os.remove(path)
+                self.logger.debug('Done removing file with path: %s' % path)
+        else:
+            return self.removeDirTree(path)
 
-    @path_check(filename={"required",})
+    @path_check(filename={"required", })
     def createEmptyFile(self, filename):
         """Create an empty file
         @param filename: string (file path name to be created)
         """
-        self.logger.debug('creating an empty file with name & path: %s' % filename)
+        self.logger.debug(
+            'creating an empty file with name & path: %s' % filename)
         open(filename, "w").close()
         self.logger.debug(
             'Empty file %s has been successfully created' % filename)
 
-    @path_check(newdir={"required",})
+    @path_check(newdir={"required", })
     def createDir(self, newdir, unlink=True):
         """Create new Directory
         @param newdir: string (Directory path/name)
@@ -215,7 +232,8 @@ class SystemFS:
                     if e.errno != os.errno.EEXIST:  # File exists
                         raise
 
-            self.logger.debug('Created the directory [%s]' % j.data.text.toStr(newdir))
+            self.logger.debug(
+                'Created the directory [%s]' % j.data.text.toStr(newdir))
 
     def copyDirTree(self, src, dst, keepsymlinks=False, deletefirst=False,
                     overwriteFiles=True, ignoredir=[".egg-info", ".dist-info"], ignorefiles=[".egg-info"], rsync=True,
@@ -274,7 +292,7 @@ class SystemFS:
                     elif self.isDir(srcname):
                         # print "1:%s %s"%(srcname,dstname)
                         self.copyDirTree(srcname, dstname, keepsymlinks, deletefirst,
-                                             overwriteFiles=overwriteFiles, applyHrdOnDestPaths=applyHrdOnDestPaths)
+                                         overwriteFiles=overwriteFiles, applyHrdOnDestPaths=applyHrdOnDestPaths)
                     else:
                         # print "2:%s %s"%(srcname,dstname)
                         self.copyFile(
@@ -321,7 +339,7 @@ class SystemFS:
 
             return j.tools.executorLocal.execute(cmd)[1]
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def removeDirTree(self, path, onlyLogWarningOnRemoveError=False):
         """Recursively delete a directory tree.
             @param path: string (the path to be removed)
@@ -355,7 +373,8 @@ class SystemFS:
         """
         self.logger.debug('Removing the directory with path: %s' % path)
         os.rmdir(path)
-        self.logger.debug('Directory with path: %s is successfully removed' % path)
+        self.logger.debug(
+            'Directory with path: %s is successfully removed' % path)
 
     @path_check(path={"required", "exists", "dir"})
     def changeDir(self, path):
@@ -377,7 +396,8 @@ class SystemFS:
         """
         self.logger.debug('Moving directory from %s to %s' % (source, destin))
         self.move(source, destin)
-        self.logger.debug('Directory is successfully moved from %s to %s' % (source, destin))
+        self.logger.debug(
+            'Directory is successfully moved from %s to %s' % (source, destin))
 
     def joinPaths(self, *args):
         """Join one or more path components.
@@ -403,7 +423,7 @@ class SystemFS:
             args = args2
         return os.path.join(*args)
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def getDirName(self, path, lastOnly=False, levelsUp=None):
         """
         Return a directory name from pathname path.
@@ -431,13 +451,13 @@ class SystemFS:
                     "Cannot find part of dir %s levels up, path %s is not long enough" % (levelsUp, path))
         return dname + os.sep if dname else dname
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def getBaseName(self, path):
         """Return the base name of pathname path."""
         self.logger.debug('Get basename for path: %s' % path)
         return os.path.basename(path.rstrip(os.path.sep))
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def pathShorten(self, path):
         """
         Clean path (change /var/www/../lib to /var/lib). On Windows, if the
@@ -459,7 +479,7 @@ class SystemFS:
                 cleanedPath = "%s%s" % (cleanedPath, sep)
         return cleanedPath
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def pathClean(self, path):
         """
         goal is to get a equal representation in / & \ in relation to os.sep
@@ -472,7 +492,7 @@ class SystemFS:
         path = path.strip()
         return path
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def pathDirClean(self, path):
         path = path + os.sep
         return self.pathClean(path)
@@ -480,7 +500,7 @@ class SystemFS:
     def dirEqual(self, path1, path2):
         return self.pathDirClean(path1) == self.pathDirClean(path2)
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def pathNormalize(self, path, root=""):
         """
         paths are made absolute & made sure they are in line with os.sep
@@ -534,7 +554,7 @@ class SystemFS:
                 result.append(item)
         return "/".join(result)
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def getParent(self, path):
         """
         Returns the parent of the path:
@@ -549,12 +569,12 @@ class SystemFS:
             return os.sep
         return os.sep.join(parts)
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def getFileExtension(self, path):
         ext = os.path.splitext(path)[1]
         return ext.strip('.')
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def chown(self, path, user, group=None):
         from pwd import getpwnam
         from grp import getgrnam
@@ -673,7 +693,7 @@ class SystemFS:
         self.logger.debug('Get current working directory')
         return os.getcwd()
 
-    @path_check(path={"required",})
+    # DO NOT USE DECORATOR, will give bugs for symlink !
     def readLink(self, path, absolute=True):
         """Works only for unix
         Return a string representing the path to which the symbolic link points.
@@ -691,7 +711,7 @@ class SystemFS:
             raise RuntimeError("cant read link, dont understand platform")
 
         if res.startswith("..") and absolute:
-            if self.isFile(path):
+            if os.path.isfile(path):
                 srcDir = self.getParent(self.getDirName(path))
             else:
                 srcDir = self.getDirName(path)
@@ -931,17 +951,19 @@ class SystemFS:
         if path is None:
             raise TypeError('Path is not passed in system.fs.exists')
         if os.path.exists(path) or os.path.islink(path):
-            if self.isLink(path) and followlinks:
-                self.logger.debug('path %s exists' % str(path.encode("utf-8")))
+            if os.path.islink(path) and followlinks:
+                self.logger.debug('path %s exists' %
+                                  str(path.encode("utf-8")))
                 relativelink = self.readLink(path)
-                newpath = self.joinPaths(self.getParent(path), relativelink)
+                newpath = self.joinPaths(
+                    self.getParent(path), relativelink)
                 return self.exists(newpath)
             else:
                 return True
         self.logger.debug('path %s does not exist' % str(path.encode("utf-8")))
         return False
 
-    @path_check(path={"required", "exists"}, target={"required",})
+    @path_check(path={"required", "exists"}, target={"required", })
     def symlink(self, path, target, overwriteTarget=False):
         """Create a symbolic link
         @param path: source path desired to create a symbolic link for
@@ -961,6 +983,9 @@ class SystemFS:
                 self.removeDirTree(target)
             else:
                 self.remove(target)
+
+        if os.path.islink(target):
+            self.remove(target)
 
         dir = self.getDirName(target)
         if not self.exists(dir):
@@ -1004,14 +1029,15 @@ class SystemFS:
         @rtype: concatenation of dirname, and optionally linkname, etc.
         with exactly one directory separator (os.sep) inserted between components, unless path2 is empty
         """
-        self.logger.debug('Create a hard link pointing to %s named %s' % (source, destin))
+        self.logger.debug(
+            'Create a hard link pointing to %s named %s' % (source, destin))
         if j.core.platformtype.myplatform.isUnix or j.core.platformtype.myplatform.isMac:
             return os.link(source, destin)
         else:
             raise j.exceptions.RuntimeError(
                 'Cannot create a hard link on windows')
 
-    @path_check(path={"required",})
+    @path_check(path={"required", })
     def checkDirParam(self, path):
         if(path.strip() == ""):
             raise TypeError("path parameter cannot be empty.")
@@ -1119,7 +1145,7 @@ class SystemFS:
         """
         return os.stat(path, follow_symlinks=follow_symlinks)
 
-    @path_check(dirname={"required", "exists", "dir"}, newname={"required",})
+    @path_check(dirname={"required", "exists", "dir"}, newname={"required", })
     def renameDir(self, dirname, newname, overwrite=False):
         """Rename Directory from dirname to newname
         @param dirname: string (Directory original name)
@@ -1338,7 +1364,8 @@ class SystemFS:
         """
         create a tmp dir name and makes sure the dir exists
         """
-        tmpdir = self.joinPaths(j.dirs.TMPDIR, j.data.idgenerator.generateXCharID(10))
+        tmpdir = self.joinPaths(
+            j.dirs.TMPDIR, j.data.idgenerator.generateXCharID(10))
         if create is True:
             self.createDir(tmpdir)
         return tmpdir
@@ -1369,7 +1396,7 @@ class SystemFS:
         """
         if dir is None:
             return self.joinPaths(j.dirs.TMPDIR, prefix +
-                                      str(j.data.idgenerator.generateRandomInt(0, 1000000000000)) + ".tmp")
+                                  str(j.data.idgenerator.generateRandomInt(0, 1000000000000)) + ".tmp")
         else:
             dir = dir or j.dirs.TMPDIR
             return tempfile.mktemp('', prefix, dir)
